requirements.txt
psycopg2==2.9.9  # duplicates build/Dockerfile
aws-wsgi==0.2.7
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
Flask-SQLAlchemy==3.1.1
Flask-cors==4.0.0
connexion==2.14.2
sentry-sdk[flask]==1.44.0
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
